Bots/store-expense-bot/requirements.txt
python-telegram-bot[job-queue]==22.6
gspread==6.1.2
google-auth==2.29.0
python-dotenv==1.0.1
pytz==2024.1
